rename #4: 파일, 폴더 이름과 경로 수정
- 라우트 경로와 페이지의 타이틀을 연상시키는 이름으로 변경함
- To 메세지를 작성하는 컴포넌트 : PostToPage
- From 메세지를 작성하는 컴포넌트 : PostFromPage
- CustomDropdown 컴포넌트는 components/create-message-form 폴더를 만들어서 하위로 들어감으로 다른 페이지에서 사용하는 컴포넌트와 구분
