+++ src/App.jsx
@@ -3,8 +3,8 @@ import './App.scss';
 import Main from './pages/Main';
 import PostedPage from './pages/PostedPage';
 import List from './pages/List';
-import MessageForm from './pages/MessageForm';
-import MessageBackgroundForm from './pages/MessageBackgroundForm';
+import PostFromPage from './pages/PostFromPage';
+import PostToPage from './pages/PostToPage';
 
 function App() {
   return (
@@ -12,8 +12,8 @@ function App() {
       <Route path='/' element={<Main />} />
       <Route path='/posted-page' element={<PostedPage />} />
       <Route path='/list' element={<List />} />
-      <Route path='/post/:id/message' element={<MessageForm />} />
-      <Route path='/post' element={<MessageBackgroundForm />} />
+      <Route path='/post/:id/message' element={<PostFromPage />} />
+      <Route path='/post' element={<PostToPage />} />
     </Routes>
   );
 }
