+++ src/styles/create-message-form/PostFromPage.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../variables';
 
 // 모바일 360~ 767 태블릿 767~1199 데스크톱 1200~
 @mixin flex-center {
